+++ agentpm/tests/lm_widgets/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for LM Indicator Widget adapter."""